frontend/requirements.txt
# Streamlit and UI
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0

# HTTP client for API calls
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
